+++ CSE/declaration_public.vb
@@ -13,7 +13,7 @@ Module declaration_public
 
     ' Description of the form
     Public Const AppName As String = "Air Drag"                 ' Name of the programm
-    Public Const AppVers As String = "2.0.7-beta9"              ' Version of the Programm
+    Public Const AppVers As String = "2.0.7-beta10"             ' Version of the Programm
     Public AppDate As String                                    ' Date of the compilation of the programm
 
     ' Control variables
@@ -50,7 +50,7 @@ Module declaration_public
 
     ' Constances for the array declaration
     Public JumpPoint As List(Of Integer)                        ' Point at that a jump in the time-resolved data is detected
-    Public OptPar() As Boolean = ({True})                       ' Array to identify if optional parameters are given
+    Public OptPar() As Boolean = ({True, True})                 ' Array to identify if optional parameters are given
     Public AT As Boolean = False                                ' Calculation of an automatic transmission (with n_card)
     Public MT_AMT As Boolean = False                            ' Calculation of an manual transmission (always when n_eng available)
     Public KoordSys() As Boolean = {False, False}               ' Used Koordinate system 0. (MM.MM); 1.(DD.DD)