+++ CHANGES.md
@@ -1,10 +1,11 @@
 VECTO-CSE: Changes
 ===================
-#### 2016-11-09: v2.0.7-beta8/9 ####
+#### 2016-11-09: v2.0.7-beta8/9/10 ####
 TUG improvements:
 
   * Correction of rolling resistance criteria
   * Correction of digit test for coordinates
+  * Update of needed signals in calibration run (t_ground no longer needed as input)
 
 #### 2016-10-20: v2.0.7-beta7 ####
 TUG improvements: